Wave 3: 199 more real names, +9 exact, zero regressions
Board 4635 -> 4644 exact, ZERO true regressions, fake-match smells 12 -> 9.

THE DEFINITIVE NAME ORACLE was found this wave: the symbol atlas contains OUR
OWN build's linker map (2002-01-14_2002_debug__cachebeta_exe), keyed by VA and
tagged with the owning .obj. It resolved every placeholder in several objects
and CORRECTED THREE NAMES earlier waves had invented -
circular_queue_validate -> circular_queue_verify, lra_block_verify ->
verify_lra_cache_block, lra_verify -> verify_lra_cache, lruv_verify ->
lruv_cache_verify. It lists statics too, so presence is authoritative while
absence proves nothing. Helper: scratch/o2_atlas.py.

199 renames applied (180 static), propagated into the dependent manifests.
Nearly all RECOVERED from the January map, the 2001-09-25 maps, or the HCEA
PDB; the few INVENTED ones are marked as such in their proposal evidence.
Highlights: public_key_crypt 6/6 with x_exp_y_mod_n and the two
Diffie-Hellman key generators; thread_win32's get_mutex_from_pool (present in
all twelve atlas builds); render_debug's eleven; the shader_transparent_generic
preprocessor's nine.

The no-fake-matching gate held: an agent found a lever reproducing January
exactly for primegen (lcsd 64) but which required moving a statement to an
unnatural place, and correctly declined to land it.

+++ source/networking/network_messages.h
@@ -82,6 +82,6 @@ boolean decode_network_game_message(
 	short *encoded_message_size,
 	short *packet_type,
 	short *packet_version,
-	long expected_packet_class);
+	short expected_packet_class);
 
 #endif // __NETWORK_MESSAGES_H
